socialaggregator/requirements.txt
Django
gunicorn
dj-database-url
dj-static
django-appconf
django-compressor
djangorestframework
drf-nested-routers
six
static3
wsgiref
django-environ

rest-social-auth
django-allauth
django-rest-auth
factory_boy
nose
django_nose
django_extensions
HTTPretty
